+++ lib/src/controller.dart
@@ -1,7 +1,8 @@
 
+import 'package:n_image_picker/src/image_viewer_dialog.dart';
+
 import 'io_file.dart' if (dart.library.web) 'web_file.dart';
 import 'dart:io' if (dart.library.web) 'dart:html';
-// import '' if (dart.library.html) 'package:n_image_picker/src/set_web.dart';
 
 import 'package:file_picker/file_picker.dart';
 import 'package:flutter/foundation.dart';
@@ -11,6 +12,7 @@ import 'package:http_parser/http_parser.dart';
 
 class NImagePickerController with ChangeNotifier{
   PlatformFile? _file;
+  Uint8List   ? _bytes;
   String        _imageKey      = "image";
   List<String>  _fileTypes     = const [ 'png', 'jpg', 'jpeg' ];
   String        _extension     = '';
@@ -28,6 +30,7 @@ class NImagePickerController with ChangeNotifier{
     "crossOrigin"                       : "Anonymous",
   };
 
+  /// Map for headers, this need a backend open port for your domain
   set headers(Map<String, String> headers) {
     _headers = headers;
     notifyListeners();
@@ -38,13 +41,14 @@ class NImagePickerController with ChangeNotifier{
     notifyListeners();
   }
 
-  ///Image name for [MultipartFile], is the same key
-  ///for posting like {"imageKey": image}
+  /// json key for posting [MultipartFile] image, example:
+  /// - "key" : "server/path/image.png"
   set imageKey(String name) {
     _imageKey = name;
     notifyListeners();
   }
 
+  /// List of supported formats
   set fileTypes(List<String> fl) {
     _fileTypes = fl;
     notifyListeners();
@@ -55,14 +59,22 @@ class NImagePickerController with ChangeNotifier{
     notifyListeners();
   }
 
+  /// Map for headers, this need a backend open port for your domain
   Map<String, String> get headers     =>  _headers;
+
+  /// json key for posting [MultipartFile] image, example:
+  /// - "key" : "server/path/image.png"
   String              get imageKey    =>  _imageKey;
-  PlatformFile?       get file        =>  _file;
+
+  /// List of supported formats
   List<String>        get fileTypes   =>  _fileTypes;
-  String              get path        =>  _file?.path   ??  '';
-  Uint8List           get bytes       =>  _file?.bytes  ??  Uint8List(0);
+
+  PlatformFile?       get file        =>  _file;
+  String              get path        =>  _file?.path ??  '';
+  Uint8List?          get bytes       =>  _bytes;
   bool                get error       =>  _error;
   bool                get hasImage    =>  _hasImage;
+  bool                get hasnoImage  =>  !_hasImage;
   bool                get fromLoading =>  _fromLoading;
   Image               get image       =>  _file == null
   ? throw Exception()
@@ -72,18 +84,20 @@ class NImagePickerController with ChangeNotifier{
     : File(_file!.path!)
   );
 
-
+  /// Set the image file from http response and url
   Future<void> setFromResponse({required Response response, required String url}) async {
     try{
       if(kIsWeb){
         await setFile(response, url, _imageKey, headers).then((r) {
           _file   = r.platformFile;
+          _bytes  = r.platformFile.bytes;
           _error  = r.error;
           notifyListeners();
         });
       } else {
         await setFile(response, url, _imageKey, headers).then((r) {
           _file     = r.platformFile;
+          _bytes    = r.platformFile.bytes;
           _error    = r.error;
           _hasImage = !r.error;
           notifyListeners();
@@ -91,34 +105,44 @@ class NImagePickerController with ChangeNotifier{
       }
     } catch (e){
       debugPrint('n_image_piker e1: $e');
+      _file     = null;
+      _bytes    = null;
       _error    = true;
       _hasImage = false;
       notifyListeners();
     }
   }
 
-  ///this only works in
+  /// This dont work in web!
   Future<void> setFromPath({required String path}) async {
-    try{
-      await setFileFromPath(path).then((r) {
-        _file     = r.platformFile;
-        _error    = r.error;
-        _hasImage = !r.error;
+    if(kIsWeb){
+      throw Exception('This dont work in web');
+    } else {
+      try{
+        await setFileFromPath(path).then((r) {
+          _file     = r.platformFile;
+          _bytes    = _file?.bytes;
+          _error    = r.error;
+          _hasImage = !r.error;
+          notifyListeners();
+        });
+      } catch (e){
+        debugPrint('n_image_piker e1: $e');
+        _file     = null;
+        _bytes    = null;
+        _error    = true;
+        _hasImage = false;
         notifyListeners();
-      });
-    } catch (e){
-      debugPrint('n_image_piker e1: $e');
-      _error    = true;
-      _hasImage = false;
-      notifyListeners();
+      }
     }
   }
 
-  ///- this will be used for get a posting image
-  ///- needs a key name for posting or key must be empty
+  /// return an async image for uploading using [imageKey] example:
+  /// - await controller.multipartFile.then((image) => image)
+  /// it works in web and platforms
   Future<MultipartFile> get multipartFile async {
     return _file == null
-    ? throw Exception()
+    ? throw Exception('There is no any image loaded')
     : kIsWeb
       ? MultipartFile.fromBytes(
         _imageKey,
@@ -133,6 +157,7 @@ class NImagePickerController with ChangeNotifier{
       );
   }
 
+  /// Open the image dialog picker
   Future<void> pickImage() async => await FilePicker.platform.pickFiles(
     type              : FileType.custom,
     allowedExtensions : _fileTypes,
@@ -140,6 +165,7 @@ class NImagePickerController with ChangeNotifier{
   ).then((response) async {
     if (response != null) {
       _file       = response.files.single;
+      _bytes      = response.files.single.bytes;
       _extension  = _file!.extension??'';
       _hasImage   = true;
     }
@@ -149,7 +175,21 @@ class NImagePickerController with ChangeNotifier{
   removeImage({required bool notify}) {
     if(!kIsWeb) if(_fromLoading) if(_file != null) if(_file!.path != null) File(_file!.path!).delete();
     _file     = null;
+    _bytes    = null;
     _hasImage = false;
+    _error    = false;
     if(notify) notifyListeners();
   }
+
+  showImageViewer(BuildContext context, {
+    required bool blur,
+    required double sigma
+  }) => _bytes == null
+  ? throw Exception('There is no any image loaded')
+  : imageViewerDialog(
+    context,
+    bytes : _bytes ?? Uint8List(0),
+    blur  : blur,
+    sigma : sigma
+  );
 }